requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements-dev.ini
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
beautifulsoup4==4.14.2
    # via yfinance
black==25.11.0
    # via -r requirements-dev.ini
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via black
colorama==0.4.6
    # via click
curl-cffi==0.13.0
    # via yfinance
flake8==7.0.0
    # via pyproject-flake8
frozendict==2.4.7
    # via yfinance
google-auth==2.43.0
    # via google-genai
google-genai==1.52.0
    # via -r requirements.ini
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
isort==7.0.0
    # via -r requirements-dev.ini
mccabe==0.7.0
    # via flake8
multitasking==0.0.12
    # via yfinance
mypy==1.18.2
    # via -r requirements-dev.ini
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
numpy==2.3.5
    # via
    #   pandas
    #   yfinance
packaging==25.0
    # via black
pandas==2.3.3
    # via yfinance
pathspec==0.12.1
    # via
    #   black
    #   mypy
peewee==3.18.3
    # via yfinance
platformdirs==4.5.0
    # via
    #   black
    #   yfinance
protobuf==6.33.1
    # via yfinance
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.11.1
    # via flake8
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.2.0
    # via flake8
pyproject-flake8==7.0.0
    # via -r requirements-dev.ini
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.ini
pytokens==0.3.0
    # via black
pytz==2025.2
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.3
    # via -r requirements.ini
requests==2.32.5
    # via
    #   google-genai
    #   yfinance
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   google-genai
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via
    #   google-genai
    #   yfinance
yfinance==0.2.66
    # via -r requirements.ini
